Rename to data.js
